Replaced VBD specific code from xen to disk directory
