+++ Makefile.am
@@ -17,7 +17,8 @@ h_private = \
     libvmi/os/os_interface.h \
     libvmi/driver/driver_interface.h \
     libvmi/driver/driver_wrapper.h \
-    libvmi/driver/memory_cache.h
+    libvmi/driver/memory_cache.h \
+    libvmi/disk/vbd_private.h
 
 c_sources = \
     libvmi/accessors.c \
@@ -38,7 +39,8 @@ c_sources = \
     libvmi/arch/ept.c \
     libvmi/driver/driver_interface.c \
     libvmi/driver/memory_cache.c \
-    libvmi/os/os_interface.c
+    libvmi/os/os_interface.c \
+    libvmi/disk/vbd.c
 
 if ENABLE_ADDRESS_CACHE
     c_sources   += libvmi/cache.c
@@ -91,8 +93,6 @@ if WITH_XEN
     h_public    += libvmi/events.h
     drivers     += libvmi/driver/xen/altp2m.c \
                    libvmi/driver/xen/altp2m_private.h \
-                   libvmi/driver/xen/vbd.c \
-                   libvmi/driver/xen/vbd_private.h \
                    libvmi/driver/xen/xen.h \
                    libvmi/driver/xen/xen_private.h \
                    libvmi/driver/xen/xen.c \
