+++ libvmi/driver/xen/CMakeLists.txt
@@ -1,8 +1,6 @@
 target_sources(vmi_shared PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/altp2m.c
     ${CMAKE_CURRENT_SOURCE_DIR}/altp2m_private.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/vbd.c
-    ${CMAKE_CURRENT_SOURCE_DIR}/vbd_private.h
     ${CMAKE_CURRENT_SOURCE_DIR}/xen.h
     ${CMAKE_CURRENT_SOURCE_DIR}/xen_private.h
     ${CMAKE_CURRENT_SOURCE_DIR}/xen.c
